refactor: Extract Node to its own file
