+++ apps/client/src/features/page/tree/components/node/node.tsx
@@ -1,0 +1,317 @@
+import { NodeApi, NodeRendererProps, TreeApi } from "react-arborist";
+import { useNavigate, useParams } from "react-router-dom";
+import { useUpdatePageMutation } from "@/features/page/queries/page-query.ts";
+import { useAtom } from "jotai";
+import { treeDataAtom } from "@/features/page/tree/atoms/tree-data-atom.ts";
+import { useQueryEmit } from "@/features/websocket/use-query-emit.ts";
+import { SpaceTreeNode } from "@/features/page/tree/types.ts";
+import { SidebarPagesParams } from "@/features/page/types/page.types.ts";
+import { queryClient } from "@/main.tsx";
+import { getSidebarPages } from "@/features/page/services/page-service.ts";
+import {
+  appendNodeChildren,
+  buildTree,
+  updateTreeNodeIcon,
+} from "@/features/page/tree/utils";
+import { buildPageUrl } from "@/features/page/page.utils.ts";
+import clsx from "clsx";
+import classes from "@/features/page/tree/styles/tree.module.css";
+import EmojiPicker from "@/components/ui/emoji-picker.tsx";
+import {
+  IconChevronDown,
+  IconChevronRight,
+  IconDotsVertical,
+  IconFileDescription,
+  IconLink,
+  IconPlus,
+  IconPointFilled,
+  IconTrash,
+} from "@tabler/icons-react";
+import { ActionIcon, Menu, rem } from "@mantine/core";
+import { useClipboard } from "@mantine/hooks";
+import { useDeletePageModal } from "@/features/page/hooks/use-delete-page-modal.tsx";
+import { getAppUrl } from "@/lib/config.ts";
+import { notifications } from "@mantine/notifications";
+import React from "react";
+
+export function Node({
+  node,
+  style,
+  dragHandle,
+  tree,
+}: NodeRendererProps<any>) {
+  const navigate = useNavigate();
+  const updatePageMutation = useUpdatePageMutation();
+  const [treeData, setTreeData] = useAtom(treeDataAtom);
+  const emit = useQueryEmit();
+  const { spaceSlug } = useParams();
+
+  async function handleLoadChildren(node: NodeApi<SpaceTreeNode>) {
+    if (!node.data.hasChildren) return;
+    if (node.data.children && node.data.children.length > 0) {
+      return;
+    }
+
+    try {
+      const params: SidebarPagesParams = {
+        pageId: node.data.id,
+        spaceId: node.data.spaceId,
+      };
+
+      const newChildren = await queryClient.fetchQuery({
+        queryKey: ["sidebar-pages", params],
+        queryFn: () => getSidebarPages(params),
+        staleTime: 10 * 60 * 1000,
+      });
+
+      const childrenTree = buildTree(newChildren.items);
+
+      const updatedTreeData = appendNodeChildren(
+        treeData,
+        node.data.id,
+        childrenTree,
+      );
+
+      setTreeData(updatedTreeData);
+    } catch (error) {
+      console.error("Failed to fetch children:", error);
+    }
+  }
+
+  const handleClick = () => {
+    const pageUrl = buildPageUrl(spaceSlug, node.data.slugId, node.data.name);
+    navigate(pageUrl);
+  };
+
+  const handleUpdateNodeIcon = (nodeId: string, newIcon: string) => {
+    const updatedTree = updateTreeNodeIcon(treeData, nodeId, newIcon);
+    setTreeData(updatedTree);
+  };
+
+  const handleEmojiIconClick = (e: any) => {
+    e.preventDefault();
+    e.stopPropagation();
+  };
+
+  const handleEmojiSelect = (emoji: { native: string }) => {
+    handleUpdateNodeIcon(node.id, emoji.native);
+    updatePageMutation.mutateAsync({ pageId: node.id, icon: emoji.native });
+
+    setTimeout(() => {
+      emit({
+        operation: "updateOne",
+        entity: ["pages"],
+        id: node.id,
+        payload: { icon: emoji.native },
+      });
+    }, 50);
+  };
+
+  const handleRemoveEmoji = () => {
+    handleUpdateNodeIcon(node.id, null);
+    updatePageMutation.mutateAsync({ pageId: node.id, icon: null });
+
+    setTimeout(() => {
+      emit({
+        operation: "updateOne",
+        entity: ["pages"],
+        id: node.id,
+        payload: { icon: null },
+      });
+    }, 50);
+  };
+
+  if (
+    node.willReceiveDrop &&
+    node.isClosed &&
+    (node.children.length > 0 || node.data.hasChildren)
+  ) {
+    handleLoadChildren(node);
+    setTimeout(() => {
+      if (node.state.willReceiveDrop) {
+        node.open();
+      }
+    }, 650);
+  }
+
+  return (
+    <>
+      <div
+        style={style}
+        className={clsx(classes.node, node.state)}
+        ref={dragHandle}
+        onClick={handleClick}
+      >
+        <PageArrow node={node} onExpandTree={() => handleLoadChildren(node)} />
+
+        <div onClick={handleEmojiIconClick} style={{ marginRight: "4px" }}>
+          <EmojiPicker
+            onEmojiSelect={handleEmojiSelect}
+            icon={
+              node.data.icon ? (
+                node.data.icon
+              ) : (
+                <IconFileDescription size="18" />
+              )
+            }
+            readOnly={tree.props.disableEdit as boolean}
+            removeEmojiAction={handleRemoveEmoji}
+          />
+        </div>
+
+        <span className={classes.text}>{node.data.name || "untitled"}</span>
+
+        <div className={classes.actions}>
+          <NodeMenu node={node} treeApi={tree} />
+
+          {!tree.props.disableEdit && (
+            <CreateNode
+              node={node}
+              treeApi={tree}
+              onExpandTree={() => handleLoadChildren(node)}
+            />
+          )}
+        </div>
+      </div>
+    </>
+  );
+}
+
+interface CreateNodeProps {
+  node: NodeApi<SpaceTreeNode>;
+  treeApi: TreeApi<SpaceTreeNode>;
+  onExpandTree?: () => void;
+}
+
+function CreateNode({ node, treeApi, onExpandTree }: CreateNodeProps) {
+  function handleCreate() {
+    if (node.data.hasChildren && node.children.length === 0) {
+      node.toggle();
+      onExpandTree();
+
+      setTimeout(() => {
+        treeApi?.create({ type: "internal", parentId: node.id, index: 0 });
+      }, 500);
+    } else {
+      treeApi?.create({ type: "internal", parentId: node.id });
+    }
+  }
+
+  return (
+    <ActionIcon
+      variant="transparent"
+      c="gray"
+      onClick={(e) => {
+        e.preventDefault();
+        e.stopPropagation();
+        handleCreate();
+      }}
+    >
+      <IconPlus style={{ width: rem(20), height: rem(20) }} stroke={2} />
+    </ActionIcon>
+  );
+}
+
+interface NodeMenuProps {
+  node: NodeApi<SpaceTreeNode>;
+  treeApi: TreeApi<SpaceTreeNode>;
+}
+
+function NodeMenu({ node, treeApi }: NodeMenuProps) {
+  const clipboard = useClipboard({ timeout: 500 });
+  const { spaceSlug } = useParams();
+  const { openDeleteModal } = useDeletePageModal();
+
+  const handleCopyLink = () => {
+    const pageUrl =
+      getAppUrl() + buildPageUrl(spaceSlug, node.data.slugId, node.data.name);
+    clipboard.copy(pageUrl);
+    notifications.show({ message: "Link copied" });
+  };
+
+  return (
+    <Menu shadow="md" width={200}>
+      <Menu.Target>
+        <ActionIcon
+          variant="transparent"
+          c="gray"
+          onClick={(e) => {
+            e.preventDefault();
+            e.stopPropagation();
+          }}
+        >
+          <IconDotsVertical
+            style={{ width: rem(20), height: rem(20) }}
+            stroke={2}
+          />
+        </ActionIcon>
+      </Menu.Target>
+
+      <Menu.Dropdown>
+        <Menu.Item
+          leftSection={<IconLink style={{ width: rem(14), height: rem(14) }} />}
+          onClick={(e) => {
+            e.preventDefault();
+            e.stopPropagation();
+            handleCopyLink();
+          }}
+        >
+          Copy link
+        </Menu.Item>
+
+        {!(treeApi.props.disableEdit as boolean) && (
+          <>
+            <Menu.Divider />
+
+            <Menu.Item
+              c="red"
+              leftSection={
+                <IconTrash style={{ width: rem(14), height: rem(14) }} />
+              }
+              onClick={(e) => {
+                e.preventDefault();
+                e.stopPropagation();
+                openDeleteModal({ onConfirm: () => treeApi?.delete(node) });
+              }}
+            >
+              Delete
+            </Menu.Item>
+          </>
+        )}
+      </Menu.Dropdown>
+    </Menu>
+  );
+}
+
+interface PageArrowProps {
+  node: NodeApi<SpaceTreeNode>;
+  onExpandTree?: () => void;
+}
+
+function PageArrow({ node, onExpandTree }: PageArrowProps) {
+  return (
+    <ActionIcon
+      size={20}
+      variant="subtle"
+      c="gray"
+      onClick={(e) => {
+        e.preventDefault();
+        e.stopPropagation();
+        node.toggle();
+        onExpandTree();
+      }}
+    >
+      {node.isInternal ? (
+        node.children && (node.children.length > 0 || node.data.hasChildren) ? (
+          node.isOpen ? (
+            <IconChevronDown stroke={2} size={18} />
+          ) : (
+            <IconChevronRight stroke={2} size={18} />
+          )
+        ) : (
+          <IconPointFilled size={8} />
+        )
+      ) : null}
+    </ActionIcon>
+  );
+}
